# contracts/openapi/devpath-openapi.yaml
openapi: 3.0.3
info:
  title: DevPath Implemented API
  version: 0.11.0
  description: >
    Implemented contract for authenticated identity, repository synchronization, deterministic analysis,
    rule evaluation, Skill Matrix, career catalog, career readiness, deterministic recommendation,
    learning roadmap, and company catalog slices. Provider access tokens and unimplemented APIs are excluded.
servers:
  - url: http://localhost:8080
    description: Local backend
paths:
...
  /api/v1/users/me:
    get:
      x-api-id: API-ID-001
      summary: Retrieve the current authenticated user
      operationId: getCurrentUser
      tags: [Identity]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current internal DevPath user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserEnvelope"
        "401":
          description: No valid DevPath session or unavailable account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/users/me/profile:
    get:
      x-api-id: API-ID-002
      summary: Retrieve the current user's profile
      operationId: getUserProfile
      tags: [Identity]
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Current user's profile.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserProfileEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    patch:
      x-api-id: API-ID-003
      summary: Update the current user's profile
      operationId: updateUserProfile
      tags: [Identity]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateUserProfileRequest"}
      responses:
        "200":
          description: Updated profile.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserProfileEnvelope"}
        "400":
          description: Invalid profile input.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/users/me/connections:
    get:
      x-api-id: API-ID-004
      summary: List the current user's active provider connections
      description: Login identity alone does not imply repository-access authorization.
      operationId: listUserConnections
      tags: [Identity, Integration]
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Active provider connections owned by the current user.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ConnectedAccountListEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/users/me/preferences:
    get:
      x-api-id: API-ID-005
      summary: Retrieve current target preferences
      operationId: getUserPreferences
      tags: [Identity]
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Current career and company targets.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserPreferenceEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/users/me/preferences/career:
    put:
      x-api-id: API-ID-006
      summary: Set the current user's career target
      operationId: setCareerTarget
      tags: [Identity]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SetCareerTargetRequest"}
      responses:
        "200":
          description: Updated target preferences.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserPreferenceEnvelope"}
        "400":
          description: Unsupported career target.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/users/me/preferences/company:
    put:
      x-api-id: API-ID-007
      summary: Set the current user's company target
      operationId: setCompanyTarget
      tags: [Identity]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SetCompanyTargetRequest"}
      responses:
        "200":
          description: Updated target preferences.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserPreferenceEnvelope"}
        "400":
          description: Unsupported company target.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
  /api/v1/analyses:
    get:
      x-api-id: API-ANA-007
      summary: List completed deterministic analyses owned by the current user
      operationId: listAnalysisHistory
      tags: [Analysis]
      security: [{sessionCookie: []}]
      parameters:
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: cursor, in: query, required: false, schema: {type: string}}
      responses:
        "200":
          description: Newest-first owner-scoped immutable analysis history.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AnalysisHistoryEnvelope"}
        "400":
          description: Invalid limit or cursor.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    post:
      x-api-id: API-ANA-001
      summary: Queue deterministic analysis for an owned repository snapshot
      operationId: requestAnalysis
      tags: [Analysis]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateAnalysisRequest"}
      responses:
        "202":
          description: Existing or newly queued owner-scoped analysis job.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AnalysisJobEnvelope"}
        "400":
          description: Invalid scope, unsupported target, or repository without a ready snapshot.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository or snapshot not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "409":
          description: Concurrent analysis request conflict.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/analysis-jobs/{jobId}:
    get:
      x-api-id: API-ANA-002
      summary: Retrieve an owner-scoped deterministic analysis job
      operationId: getAnalysisJob
      tags: [Analysis]
      security: [{sessionCookie: []}]
      parameters:
        - {name: jobId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Current durable analysis job state and safe failure detail.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AnalysisJobEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Job not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/analyses/{analysisId}:
    get:
      x-api-id: API-ANA-003
      summary: Retrieve an immutable completed analysis owned by the current user
      operationId: getAnalysisResult
      tags: [Analysis]
      security: [{sessionCookie: []}]
      parameters:
        - {name: analysisId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Immutable analysis result references.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AnalysisResultEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Analysis not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/{repositoryId}/analyses:
    get:
      x-api-id: API-REP-012
      summary: List completed deterministic analyses for an owned repository
      operationId: listRepositoryAnalysisHistory
      tags: [Analysis, Repository]
      security: [{sessionCookie: []}]
      parameters:
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: cursor, in: query, required: false, schema: {type: string}}
      responses:
        "200":
          description: Newest-first immutable analysis history for the owned repository.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AnalysisHistoryEnvelope"}
        "400":
          description: Invalid limit or cursor.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DEVPATH_SESSION
      description: Opaque HttpOnly server-session identifier. It is never read by browser JavaScript.
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Server-issued CSRF token. This is not an authentication credential.
      schema:
        type: string
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Owner-scoped stable key for one repository synchronization command.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
  schemas:
    ApiMetadata:
      type: object
      additionalProperties: false
      required: [requestId, apiVersion, timestamp]
      properties:
        requestId:
          type: string
        apiVersion:
          type: string
          example: v1
        timestamp:
          type: string
          format: date-time
...
    CurrentUser:
      type: object
      additionalProperties: false
      required: [userId, displayName, status, authenticationProvider, createdAt]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
          maxLength: 120
        avatarUrl:
          type: string
          format: uri
          nullable: true
        status:
          type: string
          enum: [ACTIVE]
        authenticationProvider:
          type: string
          enum: [GITHUB]
        createdAt:
          type: string
          format: date-time
    CurrentUserEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data:
          $ref: "#/components/schemas/CurrentUser"
        metadata:
          $ref: "#/components/schemas/ApiMetadata"
    CareerStage:
      type: string
      nullable: true
      enum: [STUDENT, ENTRY_LEVEL, JUNIOR, MID_LEVEL, SENIOR]
    UserProfile:
      type: object
      additionalProperties: false
      required: [profileId, displayName, careerStage, bio, updatedAt]
      properties:
        profileId: {type: string, format: uuid}
        displayName: {type: string, minLength: 1, maxLength: 120}
        careerStage: {$ref: "#/components/schemas/CareerStage"}
        bio: {type: string, maxLength: 1000, nullable: true}
        updatedAt: {type: string, format: date-time}
    UpdateUserProfileRequest:
      type: object
      additionalProperties: false
      required: [displayName, careerStage, bio]
      properties:
        displayName: {type: string, minLength: 1, maxLength: 120}
        careerStage: {$ref: "#/components/schemas/CareerStage"}
        bio: {type: string, maxLength: 1000, nullable: true}
    UserProfileEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/UserProfile"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    UserPreference:
      type: object
      additionalProperties: false
      required: [careerId, companyId, updatedAt]
      properties:
        careerId: {type: string, maxLength: 64, nullable: true}
        companyId: {type: string, maxLength: 64, nullable: true}
        updatedAt: {type: string, format: date-time, nullable: true}
    SetCareerTargetRequest:
      type: object
      additionalProperties: false
      required: [careerId]
      properties:
        careerId: {type: string, minLength: 1, maxLength: 64}
    SetCompanyTargetRequest:
      type: object
      additionalProperties: false
      required: [companyId]
      properties:
        companyId: {type: string, minLength: 1, maxLength: 64}
    UserPreferenceEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/UserPreference"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    ConnectedAccount:
      type: object
      additionalProperties: false
      required: [connectionId, provider, status, scopes, connectedAt, expiresAt]
      properties:
        connectionId: {type: string, format: uuid}
        provider: {type: string, enum: [GITHUB]}
        status: {type: string, enum: [ACTIVE, REVOKED]}
        scopes:
          type: array
          items: {type: string}
        connectedAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time, nullable: true}
    ConnectedAccountList:
      type: object
      additionalProperties: false
      required: [connections]
      properties:
        connections:
          type: array
          items: {$ref: "#/components/schemas/ConnectedAccount"}
...
    ConnectedAccountListEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/ConnectedAccountList"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
...
    CreateAnalysisRequest:
      type: object
      additionalProperties: false
      required: [repositoryId]
      properties:
        repositoryId: {type: string, format: uuid}
        snapshotId: {type: string, format: uuid, nullable: true}
        analysisScope: {type: string, enum: [REPOSITORY_BASELINE], default: REPOSITORY_BASELINE}
        targetCareerId:
          type: string
          nullable: true
          description: Reserved contract field; non-null values are rejected until career policies are implemented.
    AnalysisJob:
      type: object
      additionalProperties: false
      required: [jobId, jobType, status, phase, progressPercent, attemptCount, maxAttempts, submittedAt, startedAt, completedAt, pollingUrl, resultResourceUrl, errorCode, errorMessage, retryable]
      properties:
        jobId: {type: string, format: uuid}
        jobType: {type: string, enum: [REPOSITORY_ANALYSIS]}
        status: {type: string, enum: [queued, running, succeeded, failed, cancelled, expired]}
        phase: {type: string, enum: [QUEUED, EVALUATING_RULES, RETRY_WAIT, COMPLETED, FAILED]}
        progressPercent: {type: integer, minimum: 0, maximum: 100}
        attemptCount: {type: integer, minimum: 0}
        maxAttempts: {type: integer, minimum: 1}
        submittedAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time, nullable: true}
        completedAt: {type: string, format: date-time, nullable: true}
        pollingUrl: {type: string}
        resultResourceUrl: {type: string, nullable: true}
        errorCode: {type: string, nullable: true}
        errorMessage: {type: string, nullable: true}
        retryable: {type: boolean}
    AnalysisJobEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/AnalysisJob"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    AnalysisResult:
      type: object
      additionalProperties: false
      required: [analysisId, repositoryId, snapshotId, evaluationId, skillMatrixId, analysisScope, currentForRepository, completedAt]
      properties:
        analysisId: {type: string, format: uuid}
        repositoryId: {type: string, format: uuid}
        snapshotId: {type: string, format: uuid}
        evaluationId: {type: string, format: uuid}
        skillMatrixId: {type: string, format: uuid}
        analysisScope: {type: string, enum: [REPOSITORY_BASELINE]}
        currentForRepository:
          type: boolean
          description: True when this is the newest completed immutable analysis for the repository.
        completedAt: {type: string, format: date-time}
    AnalysisResultEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/AnalysisResult"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    AnalysisHistoryItem:
      allOf:
        - $ref: "#/components/schemas/AnalysisResult"
        - type: object
          additionalProperties: false
          required: [repositoryFullName, overallScore, confidence, ruleSetVersion, policyVersion]
          properties:
            repositoryFullName: {type: string, minLength: 1, maxLength: 255}
            overallScore: {type: number, format: double, minimum: 0, maximum: 100}
            confidence: {type: number, format: double, minimum: 0, maximum: 100}
            ruleSetVersion: {type: string, minLength: 1, maxLength: 32}
            policyVersion: {type: string, minLength: 1, maxLength: 32}
    AnalysisHistory:
      type: object
      additionalProperties: false
      required: [analyses, limit, nextCursor, totalCount]
      properties:
        analyses:
          type: array
          items: {$ref: "#/components/schemas/AnalysisHistoryItem"}
        limit: {type: integer, minimum: 1, maximum: 100}
        nextCursor: {type: string, nullable: true}
        totalCount: {type: integer, format: int64, minimum: 0}
    AnalysisHistoryEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/AnalysisHistory"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, metadata]
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
        metadata:
          $ref: "#/components/schemas/ApiMetadata"